+++ statgpu/survival/__init__.py
@@ -10,5 +10,6 @@ Survival analysis models.
 
 from ._cox import CoxPH
 from ._cox_cv import CoxPHCV
+from ._cox_errors import CoxCandidateNumericalError
 
-__all__ = ['CoxPH', 'CoxPHCV']
+__all__ = ['CoxPH', 'CoxPHCV', 'CoxCandidateNumericalError']